+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- `shape-features` command is added to replace `features-global` and `features-local`.
+
+### Deprecated
+
+- `features-global` command is deprecated. Use `shape-features` instead.
+- `features-local` command is deprecated. Use `shape-features` instead.
+
 ## [1.0.1] - 2026-03-24
 
 ### Fixed